Avoid a potential bool arithmetic warning

+++ include/oneapi/dpl/pstl/omp/util.h
@@ -145,7 +145,7 @@ __process_chunk(const __chunk_metrics& __metrics, _Iterator __base, _Index __chu
     bool __is_larger_chunk = __chunk_index < __metrics.__n_larger_chunks;
     std::size_t __n_previous_larger_chunks = __is_larger_chunk ? __chunk_index : __metrics.__n_larger_chunks;
     _Iterator __first = __base + __chunk_index * __metrics.__chunk_size + __n_previous_larger_chunks;
-    __f(__first, __first + __metrics.__chunk_size + __is_larger_chunk);
+    __f(__first, __first + __metrics.__chunk_size + static_cast<std::size_t>(__is_larger_chunk));
 }
 
 namespace __detail
